Remove redundant casts (per clang-tidy).

+++ src/GUI/dune/BuilderList.cpp
@@ -261,8 +261,7 @@ void BuilderList::draw(Point position) {
                     if (const auto* const pConcrete = gfx->getUIGraphic(UI_StructureSizeConcrete)) {
                         const SDL_Rect srcConcrete   = {0, 0, 1 + getStructureSize(buildItem.itemID).x * 6,
                                                         1 + getStructureSize(buildItem.itemID).y * 6};
-                        const SDL_FRect destConcrete = {static_cast<float>(dest.x + 2), static_cast<float>(dest.y + 2),
-                                                        static_cast<float>(srcConcrete.w),
+                        const SDL_FRect destConcrete = {(dest.x + 2), (dest.y + 2), static_cast<float>(srcConcrete.w),
                                                         static_cast<float>(srcConcrete.h)};
                         Dune_RenderCopyF(renderer, pConcrete, &srcConcrete, &destConcrete);
                     }
